# IGNITE-24966 Sql. Improve OrderedMergePublisher performance