--- a/modules/core/src/main/java/org/apache/ignite/internal/util/subscription/OrderedMergePublisher.java
+++ b/modules/core/src/main/java/org/apache/ignite/internal/util/subscription/OrderedMergePublisher.java
@@ -357,6 +357,8 @@ public class OrderedMergePublisher<T> implements Publisher<T> {
                         // Terminal state. No need to release guard.
                         return;
                     }
+                    default:
+                        throw new IllegalStateException("Should never get here.");
                 }
 
                 this.emitted = emitted;