@AMashenkov 2025-03-31
TPC-H query SF=0.1
Before
```
Benchmark          (fsync)  (queryId)  Mode  Cnt     Score    Error  Units
TpchBenchmark.run    false          2  avgt   20   241.401 ±  0.920  ms/op
TpchBenchmark.run    false          3  avgt   20   753.072 ±  3.418  ms/op
TpchBenchmark.run    false         8v  avgt   20   888.375 ±  2.847  ms/op
TpchBenchmark.run    false         11  avgt   20   272.972 ±  0.959  ms/op
TpchBenchmark.run    false         12  avgt   20   274.144 ±  1.671  ms/op
TpchBenchmark.run    false        12v  avgt   20   265.997 ±  1.068  ms/op
TpchBenchmark.run    false         13  avgt   20   322.637 ±  0.856  ms/op
TpchBenchmark.run    false         14  avgt   20    26.950 ±  0.102  ms/op
TpchBenchmark.run    false        14v  avgt   20    25.336 ±  0.099  ms/op
TpchBenchmark.run    false         17  avgt   20  1333.625 ± 14.113  ms/op
TpchBenchmark.run    false         21  avgt   20  3327.923 ± 10.625  ms/op

```
After
```
Benchmark          (fsync)  (queryId)  Mode  Cnt     Score    Error  Units
TpchBenchmark.run    false          2  avgt   20   185.762 ± 34.175  ms/op
TpchBenchmark.run    false          3  avgt   20   525.273 ±  1.812  ms/op
TpchBenchmark.run    false         8v  avgt   20   683.165 ±  2.896  ms/op
TpchBenchmark.run    false         11  avgt   20   178.242 ±  0.493  ms/op
TpchBenchmark.run    false         12  avgt   20   213.705 ±  0.824  ms/op
TpchBenchmark.run    false        12v  avgt   20   216.543 ±  1.857  ms/op
TpchBenchmark.run    false         13  avgt   20   233.987 ±  1.651  ms/op
TpchBenchmark.run    false         14  avgt   20    21.475 ±  0.079  ms/op
TpchBenchmark.run    false        14v  avgt   20    19.921 ±  0.097  ms/op
TpchBenchmark.run    false         17  avgt   20  1048.374 ±  4.273  ms/op
TpchBenchmark.run    false         21  avgt   20  2850.101 ± 15.694  ms/op
```