@rpuch 2025-04-02
@AMashenkov it seems that the title of this PR references a wrong issue. IGNITE-24996 is about something different :)